blacklist-backend/requirements.txt
# FastAPI 后端依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography>=41.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.4
openpyxl==3.1.2
fuzzywuzzy==0.18.0
python-levenshtein==0.23.0
celery==5.3.4
redis==5.0.1
loguru==0.7.2
python-dotenv==1.0.0
alembic==1.13.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
